refactor: replace global_server_pid with is_server_ok_global for improved signal handling in client and server

+++ src/client_bonus.c
@@ -6,13 +6,13 @@
 /*   By: katakada <katakada@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/05 20:43:32 by katakada          #+#    #+#             */
-/*   Updated: 2024/12/10 23:29:35 by katakada         ###   ########.fr       */
+/*   Updated: 2024/12/11 00:56:29 by katakada         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"
 
-static volatile sig_atomic_t	global_server_pid;
+static volatile sig_atomic_t	is_server_ok_global;
 
 static void	send_message(int pid, char *message)
 {
@@ -21,25 +21,24 @@ static void	send_message(int pid, char *message)
 	int	bit;
 
 	i = 0;
+	is_server_ok_global = 1;
 	while (1)
 	{
 		bit_count = 1;
 		while (bit_count <= 8)
 		{
+			while (is_server_ok_global == -1)
+				pause();
+			if (is_server_ok_global == 0)
+				exit(1);
+			is_server_ok_global = -1;
+			usleep(10);
 			bit = (message[i] >> (8 - bit_count)) & 1;
 			if (bit == 0)
 				kill(pid, SIGUSR1);
 			else if (bit == 1)
 				kill(pid, SIGUSR2);
-			usleep(10);
-			global_server_pid = -1;
-			while (!(global_server_pid == pid || global_server_pid == 0
-					|| global_server_pid == -2))
-				pause();
-			if (global_server_pid == -2)
-				bit_count = 1;
-			else
-				bit_count++;
+			bit_count++;
 		}
 		if (message[i] == '\0')
 			break ;
@@ -49,11 +48,16 @@ static void	send_message(int pid, char *message)
 
 static void	sig_handler(int signum, siginfo_t *siginfo, void *context)
 {
+	// static int	i = 0;
 	(void)context;
+	(void)siginfo;
+	// ft_printf("Ping %d\n", i++);
+	// ft_printf("siginfo->si_pid: %d\n", siginfo->si_pid);
+	// ft_printf("signum: %d\n", signum);
 	if (signum == SIGUSR1)
-		global_server_pid = siginfo->si_pid;
+		is_server_ok_global = 1;
 	if (signum == SIGUSR2)
-		global_server_pid = -2;
+		is_server_ok_global = 0;
 	// ft_printf("Pon!\n");
 }
 
@@ -82,7 +86,8 @@ int	main(int argc, char *argv[])
 	s_action.sa_flags = SA_SIGINFO;
 	if (sigaction(SIGUSR1, &s_action, NULL) == -1)
 		exit(1);
-	global_server_pid = pid;
+	if (sigaction(SIGUSR2, &s_action, NULL) == -1)
+		exit(1);
 	send_message(pid, argv[2]);
 	return (0);
 }
